Renamed file PaginationHelpers.jsx to TableHelpers.jsx
